shutdown pr2-head too
If there is a pr2-head machine on the network, try to shut it down too.

If the machine does not exist or does not respond, the subprocess.call
will fail and return non-zero. If it *does* succeed, wait for pr2-head
to go down before shutting down c1. Strictly speaking, this is not
needed, but it delays the shutdown beeps of c1 until pr2-head shuts down
too.

For this to work, the content of c1:/root/.ssh/id_rsa.pub has to be
added to pr2-head:/root/.ssh/authorized_keys to allow the script to
run the shutdown command on the head.